netkvm: fix wrong PhysicalMediaType in INF file
The driver INF declares PhysicalMediaType = 0, but the driver
reports NdisPhysicalMedium802_3. This does not create any
problem until we report NDIS revision 6.60 and higher, but
starting from 6.60 NDIS verifies that these numbers in INF and
in the driver are the same. Current commit make the driver
avoid assumptions regarding physical media type. Instead
it retrieves the value from the registry and reports exact
value that set by the INF file.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>