Clean code unrelated to SRIOV in Protocol driver
The SRIOV protocol driver works as a bridge for communication between
the VirtIO NetKVM driver and the VF driver, and it is only exposed to
the NetKVM and VF driver.
This sample code is a general NDIS protocol driver, most part of this
sample code are not necessary for the SRIOV protocol driver.
The following files are deleted,
    recv.c
    send.c
    excallbk.c
    debug.c
    protuser.h
Big chunks of code get deleted from the following file to simplify the
base code before SRIOV live migration implementation,
    ndisbind.c
    ntdisp.c
    ndisprot.h
    macros.h

This protocol sample code is implemented with C, it has to be change to C++
to keep synchronization with NetKVM driver C++ source code. This patch
changes file name from .c to .cpp, related C to C++ work will be implemented
in next patch. Trailing whitespaces are cleaned too.

Signed-off-by: Annie Li <annie.li@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>
Acked-by: Ross Philipson <ross.philipson@oracle.com>