Support SRIOV in NotifyObject
This patch implements NotifyObject to support 2-netdev SR-IOV live
migration. In Hyper-V, VF driver and NetVSC protocol driver binds
together in 1 to 1 mode, and this is because the LowerRange of NetVSC
protocol driver is the same as the UpperRange("ndisvf") of the VF
miniport driver.

However, in KVM, a different mechanism is used for VF-PF communication,
hence different VF device ID is advertised, and this could cause the VF
driver to bind all protocol drivers which have the LowerRange "ndis5",
this isn't properiate for 2-netdev model.

In order to support 1:1 binding of 2-netdev, a NotifyObject is needed
to bind VirtIO protocol driver with VF miniport driver and unbind this
protocol driver from all other network devices. Currently this patch
only support Intel VF driver. To support another Vender's VF driver,
a new device ID should be added into array for filter.

Related documents are introduced in the last patch of this patch set.

The class CMuxNotify is rename as CNotifyObject since the MUX model does
not exist in 2-netdev model.

Signed-off-by: Annie Li <annie.li@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>
Acked-by: Ross Philipson <ross.philipson@oracle.com>