netkvm: support devices with hash report capability only
The spec defines hash report feature as independent from
the RSS feature. So, the device can support only MQ or
even have single queue and still be able to provide
packet hash. In such case the device should be configured
using different command. The format of the payload is
compatible with one of RSS configuration.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>