NetKVM device should operate properly without VF around
If the guest runs with the NetKVM device and without the matching VF device,
the Windows system crashes when user enables the disabled NetKVM device or
changes its properties. This crash is due to the lingering NetKVM node in
the UpAdaptList of the protocol driver. So the procedure is implemented to
remove the node of the disabled NetKVM device.

Signed-off-by: Annie Li <annie.li@oracle.com>
Reported-by: Venu Busireddy <venu.busireddy@oracle.com>
Reviewed-by: Venu Busireddy <venu.busireddy@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>